Add PayoutDetail to comments
