+++ src/components/Story/StoryFooter.js
@@ -1,43 +1,8 @@
 import React, { PropTypes } from 'react';
 import numeral from 'numeral';
-import _ from 'lodash';
-import { FormattedRelative } from 'react-intl';
 import { Tooltip } from 'antd';
-import { calculatePayout } from '../../helpers/steemitHelpers';
+import PayoutDetail from '../PayoutDetail';
 import './StoryFooter.less';
-
-const AmountWithLabel = ({ label, amount }) => (
-  _.isNumber(amount)
-    ? <div>{label}: {numeral(amount).format('$0,0.00')}</div>
-    : null
-);
-
-const PayoutDetail = ({ post }) => {
-  const {
-    payoutLimitHit,
-    potentialPayout,
-    promotionCost,
-    cashoutInTime,
-    isPayoutDeclined,
-    maxAcceptedPayout,
-    pastPayouts,
-    authorPayouts,
-    curatorPayouts,
-  } = calculatePayout(post);
-
-  return (<div>
-    {payoutLimitHit && <div>Payout limit reached on this post</div>}
-    <AmountWithLabel label="Potential Payout" amount={potentialPayout} />
-    <AmountWithLabel label="Promoted" amount={promotionCost} />
-    {!isPayoutDeclined && cashoutInTime &&
-      <div>Will release <FormattedRelative value={cashoutInTime} /></div>}
-    {isPayoutDeclined && <div>Declined Payout</div>}
-    <AmountWithLabel label="Max Accepted Payout" amount={maxAcceptedPayout} />
-    <AmountWithLabel label="Total Past Payouts" amount={pastPayouts} />
-    <AmountWithLabel label="Authors Payout" amount={authorPayouts} />
-    <AmountWithLabel label="Curators Payout" amount={curatorPayouts} />
-  </div>);
-};
 
 const StoryFooter = ({ post, onLikeClick, onDislikeClick, onCommentClick, onShareClick }) => {
   const payout = parseFloat(post.pending_payout_value) || parseFloat(post.total_payout_value);
